requirements.txt
pillow-heif>=0.10.0
pillow-avif-plugin>=1.4.0
Pillow>=9.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
